Added a perm leaves script for MCCToolChest-PE,
Added a perm leaves script for MCCToolChest-PE & added a filter for MCEdit to find empty containers and print a list to the console it will searh for continers that are empty and do not have loot_tables assigned to them.

My perm leaves block script changes all leaves from decaying to permanent blockstate that doiesn't decay.